requirements.txt
matplotlib==3.7.0
networkx==3.0
numpy==1.24.2
Pillow==9.3.0
Pillow==9.4.0
Pillow==9.5.0
torch==1.13.1
torch_geometric==2.3.1
transformers==4.27.0.dev0
